refactor(tile-engine): extract AutotileEngine nested types to their own header
